Reorganize the repairing
If the problematic storage was fixed, then use STORAGE_MODE_REPAIR.
This will assume that the storage is actually writable, but not
available for the outside world just yet.

If the problematic storage is really broken, and can't e written
to, use STORAGE_MODE_CRASH.
This will not touch the storage, and possible migrate the object
into a different cluster.