Implement a throttling mechanism to avoi overwhelming the storages
Say you were repairing a storage. And in the middle, you decided to
reboot the RepairStorage worker for whatever reason. This will
cause the RepairStorage worker to re-start the repair on that
storage, which means that objects that have already been repaired
will be placed into the repair queue again.

In this scenario, you will get a lot of HTTP GET requests, which
will most likely work, at a a relatively fast pace. In certain
scenarios, this will eventually lead to a thundering effect and
bring the storage servers down.

With this change you can now specify the worker's max jobs per minute
for the entire STF instance. If the number of jobs processed goes
over the pre-configured amount, then each worker will go into a
loop waiting for the counter to be reset by the drone leader.

TODO: the max_jobs_per_minute attribute should be configurable from
the admin interface